Rework according to @t20100
- Move matplotlib package to silx.gui.plot.matplotlib
- Rename MPLColormap to Colors
- Move most of content of silx.gui.plot.Colors to matplotlib.Colors
- Lazy loading of lut resources
