+++ doc/source/modules/gui/plot/dev.rst
@@ -40,7 +40,6 @@ The :class:`PlotWindow` uses additional widgets:
 The widgets also use the following miscellaneous modules:
 
 - :mod:`.Colors` to convert colors from name to RGB(A)
-- :mod:`.MPLColormap` to embed recent matplotlib colormaps: 'magma', 'inferno', 'plasma' and 'viridis'.
 - :mod:`._utils`: utility functions
 
 The :mod:`backends` package provide the implementation of the rendering used by the :class:`Plot`.
@@ -154,14 +153,6 @@ The following modules are the modules used internally by the plot package.
 .. automodule:: silx.gui.plot.ScatterMaskToolsWidget
    :members:
    :show-inheritance:
-
-:mod:`MPLColormap`
-++++++++++++++++++
-
-.. currentmodule:: silx.gui.plot.MPLColormap
-
-.. automodule:: silx.gui.plot.MPLColormap
-   :members:
 
 :mod:`PlotEvents`
 +++++++++++++++++
