+++ silx/gui/plot/matplotlib/ModestImage.py
