+++ silx/gui/plot/matplotlib/__init__.py
@@ -32,7 +32,7 @@ to the used backend.
 
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "26/10/2016"
+__date__ = "04/05/2017"
 
 
 import sys
@@ -46,7 +46,7 @@ if 'matplotlib' in sys.modules:
         'matplotlib already loaded, setting its backend may not work')
 
 
-from .. import qt
+from ... import qt
 
 import matplotlib
