+++ silx/gui/plot/setup.py
@@ -24,7 +24,7 @@
 # ###########################################################################*/
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "16/02/2016"
+__date__ = "04/05/2017"
 
 
 from numpy.distutils.misc_util import Configuration
@@ -33,6 +33,7 @@ from numpy.distutils.misc_util import Configuration
 def configuration(parent_package='', top_path=None):
     config = Configuration('plot', parent_package, top_path)
     config.add_subpackage('_utils')
+    config.add_subpackage('matplotlib')
     config.add_subpackage('backends')
     config.add_subpackage('backends.glutils')
     config.add_subpackage('items')
