+++ silx/gui/setup.py
@@ -24,7 +24,7 @@
 # ###########################################################################*/
 __authors__ = ["T. Vincent"]
 __license__ = "MIT"
-__date__ = "16/01/2017"
+__date__ = "04/05/2017"
 
 
 from numpy.distutils.misc_util import Configuration
@@ -34,7 +34,6 @@ def configuration(parent_package='', top_path=None):
     config = Configuration('gui', parent_package, top_path)
     config.add_subpackage('_glutils')
     config.add_subpackage('qt')
-    config.add_subpackage('matplotlib')
     config.add_subpackage('plot')
     config.add_subpackage('fit')
     config.add_subpackage('hdf5')
